Add playbooks for production host
This commit adds two playbooks.

The `master.yml` playbook is the playbook used when running
`ansible-playbook`. It should always import other playbooks intended to
run together at once.

The `production.yml` playbook is an example of how a playbook for our
production host might look. It uses a host group defined in #3 in the
`inventory` file. It uses some of the reusable roles introduced in #4.
Any of those roles can be isolated with the `--tags` flag to
`ansible-playbook`.

Signed-off-by: Justin W. Flory <git@jwf.io>