Add golden image base configuration (#9)
* Attempted to add users

* Fixed return characters

* Implemented changes based on feedback

* Implemented changes based on feedback

* Added handlers; Modified User creation, fixed ssh key look up

* Added handlers; Modified User creation, fixed ssh key look up

* base: Fix syntax for SSH handler

The original syntax is in the incorrect format. Some examples in documentation show options specified like this, but only if all options are on one line. Instead, you need to write it as done in this commit.

* Fix line formatting

Fixes the YAML list syntax. See previous commit for more info.

* base: Add "/usr" to default shell in user creation

It's a minor technicality, but in modern RPM distributions, the full paths are `/usr/bin` and the `/bin` path is a symbolic link.