inventory: Add production, staging, centos7, and rhel8 hosts
This commit adds the hosts on the internal network set up by @leong96.
All hosts belong to one or multiple groups depending on the type of
machine they are.

Once this commit is merged, any playbook that specifies any of these
host groups will work with the correct IP addresses.

Signed-off-by: Justin W. Flory <git@jwf.io>