playbooks: Write a CentOS 7 playbook
This commit reorganizes the playbooks. Instead of a production/staging
playbook (which isn't truly accurate yet since we aren't deploying
anything), I consolidated to a single playbook for all CentOS 7 hosts.

This can be automated as a cron job with the following command:

`ansible-playbook /path/to/rit-nssa-320-infrastructure/playbooks/master.yml`

Some experimentation with what flags are needed for key/password
authentication is likely needed too.

Signed-off-by: Justin W. Flory <git@jwf.io>