Add InfluxDB role (#41)
* roles: Install my InfluxDB role from Ansible Galaxy

This commit installs my InfluxDB role from Ansible Galaxy:

```sh
ansible-galaxy install jwflory.influxdb
```

Reference: https://github.com/jwflory/ansible-role-influxdb

Signed-off-by: Justin W. Flory <git@jwf.io>

* influxdb: Add variables to override from defaults

Signed-off-by: Justin W. Flory <git@jwf.io>

* influxdb: Comment out tasks for HTTPS

We will not run InfluxDB over HTTPS for simplicity while finishing the
final project. This is a non-compatible change with the upstream change
and will require some manual work when upgrading.

Signed-off-by: Justin W. Flory <git@jwf.io>

* tests: Fix invalid YAML

This commit fixes a YAML error. There needs to be a new line character
at the end of the file.

Signed-off-by: Justin W. Flory <git@jwf.io>

* Add playbook to install/manage InfluxDB server

The added playbook should install and configure InfluxDB on a host. It
used the `influxdb` host group in the `inventory` added by @leong96.
This is not yet tested at time of commit, but will be tested in RLES.

Signed-off-by: Justin W. Flory <git@jwf.io>